🚚 Move: `response` function now has its own file [#63]
