+++ src/client/mod.rs
@@ -1,7 +1,8 @@
 //! Client for API requests.
 
 use crate::{
-    error::{ResponseError, Status},
+    client::response::response,
+    error::ResponseError,
     model::{
         achievement_info::AchievementInfoResponse,
         labs::{
@@ -27,9 +28,7 @@ use crate::{
         user_records::UserRecordsResponse,
     },
 };
-use http::status::StatusCode;
-use reqwest::{self, Error, Response};
-use serde::Deserialize;
+use reqwest::{self};
 
 const API_URL: &str = "https://ch.tetr.io/api/";
 
@@ -1112,34 +1111,7 @@ impl Client {
     }
 }
 
-/// Receives `Result<Response, Error>` and returns `Result<T, ResponseError>`.
-///
-/// # Examples
-///
-/// ```ignore
-/// let res = self.client.get(url).send().await;
-/// response(res).await
-/// ```
-async fn response<T>(response: Result<Response, Error>) -> RspErr<T>
-where
-    for<'de> T: Deserialize<'de>,
-{
-    match response {
-        Ok(r) => {
-            if !r.status().is_success() {
-                match StatusCode::from_u16(r.status().as_u16()) {
-                    Ok(c) => return Err(ResponseError::HttpErr(Status::Valid(c))),
-                    Err(e) => return Err(ResponseError::HttpErr(Status::Invalid(e))),
-                }
-            }
-            match r.json().await {
-                Ok(m) => Ok(m),
-                Err(e) => Err(ResponseError::DeserializeErr(e.to_string())),
-            }
-        }
-        Err(e) => Err(ResponseError::RequestErr(e.to_string())),
-    }
-}
+mod response;
 
 pub mod stream {
     //! Features for streams.
